backend/requirements.txt
absl-py==2.4.0
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
astunparse==1.6.3
bcrypt==5.0.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
ecdsa==0.19.1
et_xmlfile==2.0.0
flatbuffers==25.12.19
fonttools==4.61.1
funcy==2.0
gast==0.7.0
gensim==4.4.0
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
nltk==3.9.3
numexpr==2.14.1
numpy==2.4.1
openai==2.24.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.19.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
patsy==1.0.2
pillow==12.1.0
pluggy==1.6.0
protobuf==7.34.1
pyasn1==0.6.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyLDAvis==3.4.1
PyMuPDF==1.27.2
pyparsing==3.3.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-jose==3.5.0
python-pptx==1.0.2
pytz==2025.2
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setuptools==82.0.0
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
statsmodels==0.14.6
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
wheel==0.46.3
wordcloud==1.9.6
wrapt==2.1.1
xgboost==3.1.3
xlsxwriter==3.2.9
fastapi
uvicorn
sqlalchemy
python-dotenv
